refactor: fetch tooltip data before showing
